сделал все что мог
