+++ js/fetch.js
@@ -1,13 +1,9 @@
-import { validateAllHashTags } from './pristine.js';
+import { workingErrorForm } from './mistakes.js';
 import { displayUserPhotos } from './picture.js';
 import { displayBigPhoto  } from './creation_big_picture.js';
-import { filterUserPhoto, test } from './filter.js';
-const workingErrorForm = (status) => {
-  document.querySelector('.img-upload__overlay').classList.add('hidden');
-  document.querySelector(status).classList.remove('hidden');
-};
+import { filterUserPhoto, filteringNewArray } from './filter.js';
 
-const createLoader = () => {
+export const createLoader = () => {
   fetch('https://26.javascript.pages.academy/kekstagram', {
     method: 'POST',
     body: new FormData(document.querySelector('.img-upload__form')),
@@ -24,19 +20,6 @@ const createLoader = () => {
     });
 };
 
-export const sentDate = () => {
-  document.querySelector('#upload-select-image').addEventListener('submit', (evt) => {
-    evt.preventDefault();
-    if (validateAllHashTags(document.querySelector('.text__hashtags').value) === false) {
-      workingErrorForm('#error3');
-    }
-    else {
-      document.querySelector('#upload-submit').disabled = true;
-      createLoader();
-    }
-  });
-};
-
 export const qetDate = () => {
   fetch('https://26.javascript.pages.academy/kekstagram/data')
     .then((response) => {
@@ -47,7 +30,7 @@ export const qetDate = () => {
     })
     .then((response) => response.json())
     .then((photos) => {
-      test(photos);
+      filteringNewArray(photos);
       document
         .querySelector('.pictures.container')
         .append(displayUserPhotos(filterUserPhoto(photos),
