+++ js/main.js
@@ -2,7 +2,7 @@ import { qetDate } from './fetch.js';
 import { uploadNewImage } from './upload_form.js';
 import { replacingPhotoEffects } from './effects_photo.js';
 import { createSlider } from './slider.js';
-import { sentDate } from './fetch.js';
+import { sentDate } from './mistakes.js';
 import { handlingError, clearFormAfterSubmit } from './mistakes.js';
 import { lockButtonEsc } from './close_form.js';
